configure model by model name
